Added assume yes switch and preview switch to cli

+++ cockroach.py
@@ -79,12 +79,13 @@ class ConsumerGroup(object):
     #            pass
 
 class CockRoach(object):
-    def __init__(self, zkHost, stale_max_days=30, assume_yes=False):
+    def __init__(self, zkHost, stale_max_days=30, assume_yes=False, preview=False):
         self.ConsumerGroups = []
         self.zk_client = KazooClient(hosts=zkHost)
         self.zk_client.start()
         self.stale_max_days = stale_max_days
         self.assume_yes = assume_yes
+        self.preview = preview
         if self.zk_client.exists("/consumers"):
             for cg_name in self.zk_client.get_children("/consumers"):
                 self.ConsumerGroups.append(ConsumerGroup(cg_name, \
@@ -109,10 +110,10 @@ class CockRoach(object):
         stale_cgroups = self.get_stale_cgroups(display=False)
         for stale_cg in stale_cgroups:
             print stale_cg
-            if assume_yes:
+            if self.assume_yes is False:
                 confirm = raw_input("Delete?")
             else:
-                confirm="Y"
+                confirm = "Y"
 
             if confirm == "Y":
                 self.delete_cgroup(stale_cg)
@@ -120,7 +121,12 @@ class CockRoach(object):
     def delete_cgroup(self, consumergroup):
         """Deletes a consumer Group"""
         print "Deleting %s" % (consumergroup.gid)
-        self.zk_client.delete("/consumers/%s" % (consumergroup.gid), version=-1, recursive=True)
+        if self.preview is False:
+          self.zk_client.delete("/consumers/%s" % (consumergroup.gid), version=-1, recursive=True)
+          print "executed"
+        else:
+            print "pass"
+
 
     def __str__(self):
         ret = ""
@@ -136,16 +142,17 @@ if __name__ == '__main__':
                                      "Cleanup stale consumer groups from ZooKeeper")
     argparser.add_argument('zk', help="zookeeper host", default="aquzoosys031010.c031.digitalriverws.net:2182")
     argparser.add_argument('--stale', help="Search for Stale ConsumerGroups (just print)", default=False, action='store_true')
-    argparser.add_argument('--delete', help="Delete Stale ConsumerGroups", default=False, action='store_true')
+    argparser.add_argument('--delete_stale', help="Delete Stale ConsumerGroups", default=False, action='store_true')
     argparser.add_argument('--stale_max_days', type=int, help="Define after how many days a CG is considered stale", default=30)
     argparser.add_argument('--YES', help="Assume Yes on all questions", default=False, action='store_true')
+    argparser.add_argument('--preview', help="Do not execute any modifications", default=False, action='store_true')
 
     args = argparser.parse_args()
-    cockroach = CockRoach(zkHost=args.zk, stale_max_days=args.stale_max_days,assume_yes=args.YES)
+    cockroach = CockRoach(zkHost=args.zk, stale_max_days=args.stale_max_days, assume_yes=args.YES, preview=args.preview)
 
     if args.stale:
         cockroach.get_stale_cgroups(display=True)
-    elif args.delete:
+    elif args.delete_stale:
         cockroach.delete_stale_cgroups()
     else:
         print "CockRoach: %s" % (cockroach)
